MCP_Agentic_AI_Crash_Course_With_Python/pyproject.toml
[project]
name = "mcp-agentic-ai-crash-course"
version = "0.1.0"
description = "Implementation of the Agentic AI Crash Course with Python using MCP"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
    "mcp[cli]>=0.1.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.335",
    "langchain-community>=0.0.13",
    "langchain-core>=0.1.17",
    "pydantic>=2.5.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]
